+++ src/openlmis-quantity-unit-toggle/quantity-unit.service.js
@@ -61,7 +61,9 @@
             var packs = (stockOnHand > 0) ? Math.floor(stockOnHand / netContent) : Math.ceil(stockOnHand / netContent);
             var remainderDoses = stockOnHand % netContent;
 
-            return packs + ' ( +' + remainderDoses + messageService.get('openlmisInputDosesPacks.DosesBracket');
+            var doseSign = remainderDoses < 0 ? '' : '+';
+            var bracket = messageService.get('openlmisInputDosesPacks.DosesBracket');
+            return packs + ' ( ' + doseSign + remainderDoses + bracket;
         }
 
         /**